requirements.txt
pandas
streamlit
mysql.connector
geocoder
google-api-python-client
folium
streamlit_folium